added two fingers

+++ app/src/main/java/com/obolonnyy/chartapplication/chart/ChartPressedState.kt
@@ -3,7 +3,7 @@ package com.obolonnyy.chartapplication.chart
 sealed class ChartPressedState {
     object Unpressed : ChartPressedState()
     data class PressOneFinger(val x: Float, val y: Float) : ChartPressedState()
-    data class PressTwoFingers(val x1: Int, val y1: Int, val x2: Int, val y2: Int) :
+    data class PressTwoFingers(val x1: Float, val y1: Float, val x2: Float, val y2: Float) :
         ChartPressedState()
 
 }
